#1542 split userspace NAT runtime into per-concern modules
Closes #1542.

Wave-2 modularity refactor — splits userspace-dp/src/nat.rs (1605 LOC,
single file) into a nat/ directory with one file per concern:

- nat/mod.rs (~90 LOC)         — NatDecision (cross-cutting output)
                                  and the curated pub(crate) re-export
                                  surface for crate::nat::*
- nat/allocator.rs (~590 LOC)  — PortAllocator + live/persistent lease
                                  state machine + rollback activation
                                  fields + GC + recycled-port indexes.
                                  Single Mutex<PortAllocatorLiveState>
                                  serializes every structural mutation;
                                  the relaxed atomic counters rely on
                                  the mutex for happens-before, so the
                                  entire critical section stays local
                                  to this file.
- nat/source.rs (~470 LOC)     — Source NAT rule parsing, matching,
                                  match_source_nat_result_for_tuple,
                                  and release/rollback entry points
                                  that drive the allocator.
- nat/destination.rs (~180 LOC)— DnatTable
- nat/static_nat.rs (~90 LOC)  — StaticNatTable
- nat/status.rs (~35 LOC)      — source_nat_pool_statuses
- nat/tests.rs (~2700 LOC)     — full nat test suite, relocated from
                                  the previous src/nat_tests.rs

Submodules are private (mod allocator;, not pub mod allocator;); mod.rs
re-exports the pub(crate) symbols so every external caller continues to
reach them as crate::nat::<symbol>. Cross-submodule internal items use
pub(super) and are NOT re-exported. White-box tests reach allocator
internals through a debug_live() accessor on PortAllocator plus the
pub(super) field promotions on PortAllocatorLiveState
(persistent_by_source, lease_expirations, lease_expirations_by_addr,
addr_index_by_translated, recycled_ports_by_addr).

Behavioural preservation:

- No changes to NatDecision wire shape, variants, or derive set
  (Clone, Copy, Debug, Default, PartialEq, Eq).
- No new allocations on the hot path (match_source_nat_result_for_tuple
  is unchanged byte-for-byte aside from imports).
- Allocate / release / rollback ordering preserved bit-for-bit;
  activation_saw_completion / activation_previous_expires_at_ns /
  activation_had_previous_lease fields stay together with their only
  mutating sites inside allocator.rs.
- PortAllocator public surface (new, address_index, try_next_port,
  port_low, port_high) unchanged. The internal state-machine methods
  (allocate_translation, release_flow, rollback_flow, snapshot) are
  pub(super) instead of fn-private so source.rs and status.rs can
  drive them, but their signatures are byte-identical.
- DnatTable, StaticNatTable, source_nat_pool_statuses,
  parse_source_nat_rules{_with_previous}, release_source_nat_allocation,
  rollback_source_nat_allocation all preserve their pub(crate)
  signatures at crate::nat::*.

Methodology (per #1542 plan in docs/pr/1542-nat-runtime-split/plan.md):

- Plan v1 → Codex PLAN-NEEDS-MAJOR (3 MAJORs on visibility model,
  cross-submodule API table, test reachability) / AGY PLAN-READY.
- Plan v2 addressed all 3 Codex MAJORs → Codex PLAN-NEEDS-MAJOR
  (5 mechanical fix-ups) / AGY PLAN-NEEDS-MINOR (3 named adjustments
  that overlapped Codex findings).
- Plan v3 incorporated all round-2 fix-ups; architectural premise
  ratified by both reviewers. Implementation followed v3 exactly.

Tests:

- TMPDIR=/dev/shm CARGO_TARGET_DIR=/dev/shm cargo build clean.
- cargo test --release: 1417 main + 212 nat + 46 fairness pass.
- 5x flake on `nat`: 5/5 clean.
- Go suite: 30 packages clean.
- Pre-existing snat_contract_doc_guard failure on origin/master is
  not introduced by this PR (the test asserts docs/userspace-dataplane-gaps.md
  mentions "fail-closed"; that doc has never mentioned it on master).

No per-PR smoke per Wave-2 rules — NAT correctness validated by the
every-10 batch smoke. AWAITING-BATCH-MERGE marker posted.
