+++ docs/pr/1542-nat-runtime-split/reviewer-ids.md
@@ -1,0 +1,11 @@
+# #1542 NAT runtime split — reviewer task IDs
+
+## Round 1 (PLAN review on plan v1, commit 724987e6)
+
+- Codex: `task-mpmyrnf2-1de5ha` → PLAN-NEEDS-MAJOR (3 MAJORs)
+- AGY: `adversarial-review-mpmys6pk-3ut2f2` → PLAN-READY
+
+## Round 2 (PLAN review on plan v2, commit e59c6f4e)
+
+- Codex: `task-mpmz1whj-5t3zod`
+- AGY: `adversarial-review-mpmz27os-jlg8jq`
